Monetization (#124)
* Integrate LemonSqueezy for payment processing (#113)

* handle lemonsqueezy subscriptions and create entitlements from them

* update existing entitlements

* add premium page and start implementing endpoints

* add pricing list to premium page

* fix build errors

* move plan info to backend and more refactoring

* implement subscription management

* store plan id in entitlements instead of lemonsqueezy ids

* [WIP] Entitlements (#123)

* implement mvp entitlement feature computation

* centralize feature computation

* enforce monthly credit usage limit

* enforce guild count limit

* minor

* minor

* Collaborators (#121)

* collaborators MVP

* fix rebase

* enforce max collaborators from entitlements

* minor

* minor

* Hand out roles to subscribers (#126)

* update json example

* hand out roles to premium subscribers

* run plan manager

* sqlc generate

+++ kite-service/internal/api/handler/app/handler.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/kitecloud/kite/kite-service/internal/api/handler"
 	"github.com/kitecloud/kite/kite-service/internal/api/wire"
+	"github.com/kitecloud/kite/kite-service/internal/core/plan"
 	"github.com/kitecloud/kite/kite-service/internal/model"
 	"github.com/kitecloud/kite/kite-service/internal/store"
 	"github.com/kitecloud/kite/kite-service/internal/util"
@@ -15,12 +16,21 @@ import (
 
 type AppHandler struct {
 	appStore       store.AppStore
+	userStore      store.UserStore
+	planManager    *plan.PlanManager
 	maxAppsPerUser int
 }
 
-func NewAppHandler(appStore store.AppStore, maxAppsPerUser int) *AppHandler {
+func NewAppHandler(
+	appStore store.AppStore,
+	userStore store.UserStore,
+	planManager *plan.PlanManager,
+	maxAppsPerUser int,
+) *AppHandler {
 	return &AppHandler{
 		appStore:       appStore,
+		userStore:      userStore,
+		planManager:    planManager,
 		maxAppsPerUser: maxAppsPerUser,
 	}
 }
@@ -196,6 +206,10 @@ func (h *AppHandler) HandleAppTokenUpdate(c *handler.Context, req wire.AppTokenU
 }
 
 func (h *AppHandler) HandleAppDelete(c *handler.Context) (*wire.AppDeleteResponse, error) {
+	if !c.UserAppRole.CanDeleteApp() {
+		return nil, handler.ErrForbidden("missing_permissions", "You don't have permissions to delete this app")
+	}
+
 	if err := h.appStore.DeleteApp(c.Context(), c.App.ID); err != nil {
 		return nil, fmt.Errorf("failed to delete app: %w", err)
 	}
